refactor to fetcher file, test
